clean up of oscar client
- Updated video for event preview/details
- Added file upload/removal  for adjudication in event details
- Removed unused code, components, assets
- Cleaned out the event preview/event details to adhere to file server videos
- added auth request for file server
- cleaned up report/adjudication requests
