+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "OpenSensorViewer",
-  "version": "1.1.0",
+  "name": "OSCARViewer",
+  "version": "2.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
