+++ public/index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" href="./images/opensensorhub.png" />
-    <title>OpenSensorViewer</title>
+    <link rel="icon" type="image/png" href="opensensorhub.png" />
+    <title>OSCAR Viewer</title>
 </head>
 <body style="margin: 0; overflow: hidden" oncontextmenu="return false">
 <div id="root" style="width: 100%; height: 100%"></div>
