+++ src/app/_components/dashboard/LaneStatus.tsx
@@ -76,19 +76,6 @@ export default function LaneStatus(props: {dataSourcesByLane: any, initialLanes:
 
   useEffect(() => {
     addSubscriptionCallbacks();
-
-    return() => {
-      console.log("Dashboard: Lane Status unmounted, disconnecting from datasources")
-
-      //clean up subscriptions and disconnect from datasources
-      for (let [laneName, laneDSColl] of props.dataSourcesByLane.entries()) {
-
-        laneDSColl.addDisconnectToALLDSMatchingName('connectionRT');
-        laneDSColl.addDisconnectToALLDSMatchingName('tamperRT');
-        laneDSColl.addDisconnectToALLDSMatchingName('neutronRT');
-        laneDSColl.addDisconnectToALLDSMatchingName('gammaRT');
-      }
-    };
   }, [props.dataSourcesByLane]);
 
 
@@ -166,7 +153,7 @@ export default function LaneStatus(props: {dataSourcesByLane: any, initialLanes:
             {(
                 <Grid container columns={{sm: 12, md: 24, lg: 36, xl: 48}} spacing={1}>
                   {statusList.map((item) => (
-                      <Grid item sm={8} md={8} lg={8} xl={6}>
+                      <Grid key={item.id} item sm={8} md={8} lg={8} xl={6}>
                         <div onClick={() => handleLaneView(item.name)}>
                           <LaneStatusItem
                               key={item.id}
