+++ src/app/event-details/page.tsx
@@ -4,25 +4,17 @@ import {Button, Grid, Paper, Stack, Typography} from "@mui/material";
 import React, {useCallback, useContext, useEffect, useRef, useState} from "react";
 import BackButton from "../_components/BackButton";
 import DataRow from "../_components/event-details/DataRow";
-
 import MiscTable from "../_components/event-details/MiscTable";
 import {useSelector} from "react-redux";
 import ConSysApi from "osh-js/source/core/datasource/consysapi/ConSysApi.datasource";
 import AdjudicationDetail from "@/app/_components/adjudication/AdjudicationDetail";
 import Media from "@/app/_components/event-details/Media";
 import {LaneMapEntry} from "@/lib/data/oscar/LaneCollection";
-import {selectEventPreview, setSelectedRowId} from "@/lib/state/EventPreviewSlice";
+import {selectEventPreview} from "@/lib/state/EventPreviewSlice";
 import {DataSourceContext} from "@/app/contexts/DataSourceContext";
 import {PictureAsPdfRounded} from "@mui/icons-material";
 import {useReactToPrint} from "react-to-print";
 
-/**
- * Expects the following search params:
- * startTime: string;
- * endTime: string;
- *
- * Need to implement an error page to handle invalid/no search params
- */
 
 export default function EventDetailsPage() {
 
@@ -33,9 +25,7 @@ export default function EventDetailsPage() {
 
     const [gammaDatasources, setGammaDS] = useState<typeof ConSysApi[]>([]);
     const [neutronDatasources, setNeutronDS] = useState<typeof ConSysApi[]>([]);
-    const [occDatasources, setOccDS] = useState<typeof ConSysApi[]>([]);
     const [thresholdDatasources, setThresholdDS] = useState<typeof ConSysApi[]>([]);
-    const [videoDatasources, setVideoDatasources] = useState<typeof ConSysApi[]>([]);
 
     const contentRef = useRef<HTMLDivElement>(null);
     const docTitle = eventPreview.eventData ? `eventdetails-${eventPreview.eventData.laneId}-${eventPreview.eventData.observationId}-${eventPreview.eventData.startTime}-${eventPreview.eventData.endTime}` : 'eventdetails';
@@ -59,7 +49,6 @@ export default function EventDetailsPage() {
         setLocalDSMap(datasources);
         tempDSMap = datasources;
 
-
         if(!tempDSMap){
             return;
         }
@@ -67,26 +56,19 @@ export default function EventDetailsPage() {
         const updatedGamma = tempDSMap.get("gamma") || [];
         const updatedNeutron = tempDSMap.get("neutron") || [];
         const updatedThreshold = tempDSMap.get("gammaTrshld") || [];
-        const updatedVideo = tempDSMap.get("video") || [];
-        const updatedOcc = tempDSMap.get("occ") || [];
 
         setGammaDS(updatedGamma);
         setNeutronDS(updatedNeutron);
         setThresholdDS(updatedThreshold);
-        setVideoDatasources(updatedVideo);
-        setOccDS(updatedOcc);
         setDatasourcesReady(true);
-
 
     }, [eventPreview, laneMapRef]);
 
 
     useEffect(() => {
-
         async function callCollectDatasources(){
             await collectDataSources();
         }
-
 
         if(laneMapRef.current && eventPreview) {
             callCollectDatasources();
@@ -129,16 +111,15 @@ export default function EventDetailsPage() {
             </Paper>
 
             { (gammaDatasources.length > 0 || neutronDatasources.length > 0 || thresholdDatasources.length > 0) && laneMapRef &&
-
-                <Media eventData={eventPreview.eventData}  datasources={{
-                    gamma: gammaDatasources?.[0],
-                    neutron: neutronDatasources?.[0],
-                    threshold: thresholdDatasources?.[0],
-                    video: videoDatasources
+                <Media
+                    eventData={eventPreview.eventData}
+                    datasources={{
+                        gamma: gammaDatasources?.[0],
+                        neutron: neutronDatasources?.[0],
+                        threshold: thresholdDatasources?.[0],
                 }}
-                       laneMap={laneMapRef.current}
+                    laneMap={laneMapRef.current}
                 />
-
             }
 
             <Paper variant='outlined' sx={{width: "100%"}}>
