--- src/lib/data/oscar/OSCARConstants.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2024.  Botts Innovative Research, Inc.
- * All Rights Reserved
- */
-
-enum ObservationRootTypes {
-    VIDEO = "video",
-    GAMMA_COUNT = "Gamma Count",
-
-}
