+++ src/lib/data/oscar/adjudication/models/AdjudicationConstants.ts
@@ -3,108 +3,6 @@
  * All Rights Reserved
  */
 
-export const AdjudicationDatastreamConstant: any =
-    {
-    "name": "Occupancy Adjudication",
-    "outputName": "Adjudication",
-    "schema": {
-        "obsFormat": "application/swe+json",
-        "recordSchema": {
-            "type": "DataRecord",
-            "label": "Occupancy Adjudication Record",
-            "fields": [
-                {
-                    "type": "Time",
-                    "label": "Sampling Time",
-                    "name": "time",
-                    "definition": "http://www.opengis.net/def/property/OGC/0/SamplingTime",
-                    "referenceFrame": "http://www.opengis.net/def/trs/BIPM/0/UTC",
-                    "uom": {
-                        "href": "http://www.opengis.net/d…uom/ISO-8601/0/Gregorian"
-                    }
-                },
-                {
-                    "type": "Text",
-                    "name": "username",
-                    "definition": "http://sensorml.com/ont/swe/property/Username",
-                    "label": "Username"
-                },
-                {
-                    "type": "Text",
-                    "name": "feedback",
-                    "definition": "http://sensorml.com/ont/swe/property/Feedback",
-                    "label": "Feedback"
-                },
-                {
-                    "type": "Category",
-                    "name": "adjudicationCode",
-                    "definition": "http://sensorml.com/ont/swe/property/AdjudicationCode",
-                    "label": "Adjudication Code",
-                    "constraint": {
-                        "values": [
-                            "Code 1: Contraband Found",
-                            "Code 2: Other",
-                            "Code 3: Medical Isotope Found",
-                            "Code 4: NORM Found",
-                            "Code 5: Declared Shipment of Radioactive Material",
-                            "Code 6: Physical Inspection Negative",
-                            "Code 7: RIID/ASP Indicates Background Only",
-                            "Code 8: Other",
-                            "Code 9: Authorized Test, Maintenance, or Training Activity",
-                            "Code 10: Unauthorized Activity",
-                            "Code 11: Other",
-                            ""
-                        ]
-                    }
-                },
-                {
-                    "type": "Text",
-                    "name": "isotopes",
-                    "definition": "http://sensorml.com/ont/swe/property/Istotopes",
-                    "label": "Isotopes"
-                },
-                {
-                    "type": "Category",
-                    "name": "secondaryInspectionStatus",
-                    "definition": "http://sensorml.com/ont/swe/property/SecondaryInspectionStatus",
-                    "label": "Secondary Inspection Status",
-                    "constraint": {
-                        "values": [
-                            "NONE",
-                            "REQUESTED",
-                            "COMPLETE"
-                        ]
-                    }
-                },
-                {
-                    "type": "Text",
-                    "name": "filePaths",
-                    "definition": "http://sensorml.com/ont/swe/property/FilePaths",
-                    "label": "Supplemental File Paths"
-                },
-                {
-                    "type": "Text",
-                    "name": "occupancyId",
-                    "definition": "http://sensorml.com/ont/swe/property/OccupancyID",
-                    "label": "Occupancy ID"
-                },
-                {
-                    "type": "Text",
-                    "name": "alarmingSystemUid",
-                    "definition": "http://sensorml.com/ont/swe/property/SystemUID",
-                    "label": "UID of Alarming System"
-                },
-                {
-                    "type": "Text",
-                    "name": "vehicleId",
-                    "definition": "http://sensorml.com/ont/swe/property/VehicleIdentification",
-                    "label": "Vehicle ID"
-                }
-            ]
-        }
-    }
-}
-
 export interface AdjudicationCode {
     code: number;
     label: string;
