+++ types/new-types.d.ts
@@ -22,6 +22,7 @@ export interface IEventTableData {
   adjudicatedCode?: number; // Adjudication code for event
   isAdjudicated?: boolean;
   foiId: string;
+  videoFiles: string[]
 }
 
 export interface INationalTableData {
